kuhackathonbackend/pyproject.toml
[tool.poetry]
name = "kuhackathonbackend"
version = "0.1.0"
description = ""
authors = ["ratchaphon1412 <ratchaphon.h111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.7"
django-filter = "^23.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
origin = "^1.0.6"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
gunicorn = "^21.2.0"
django-seed = "^0.3.1"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.1"
typing-extensions = "^4.8.0"
django-rest-swagger = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
